Use ParquetReadSupportWrapper instead of OapReadSupportImpl
