+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/io/ParquetDataFile.scala
@@ -28,7 +28,7 @@ import org.apache.parquet.hadoop.api.RecordReader
 import org.apache.spark.sql.catalyst.InternalRow
 import org.apache.spark.sql.catalyst.expressions.UnsafeRow
 import org.apache.spark.sql.execution.datasources.oap.filecache._
-import org.apache.spark.sql.execution.datasources.parquet.ParquetReadSupportHelper
+import org.apache.spark.sql.execution.datasources.parquet.ParquetReadSupportWrapper
 import org.apache.spark.sql.types.StructType
 
 
@@ -48,7 +48,7 @@ private[oap] case class ParquetDataFile(
     val meta: ParquetDataFileHandle = DataFileHandleCacheManager(this)
 
     initRecordReader(
-      new DefaultRecordReader[UnsafeRow](new OapReadSupportImpl,
+      new DefaultRecordReader[UnsafeRow](new ParquetReadSupportWrapper,
         file, configuration, meta.footer))
   }
 
@@ -63,7 +63,7 @@ private[oap] case class ParquetDataFile(
       val meta: ParquetDataFileHandle = DataFileHandleCacheManager(this)
 
       initRecordReader(
-        new OapRecordReader[UnsafeRow](new OapReadSupportImpl,
+        new OapRecordReader[UnsafeRow](new ParquetReadSupportWrapper,
           file, configuration, rowIds, meta.footer))
     }
   }
@@ -84,7 +84,7 @@ private[oap] case class ParquetDataFile(
       }
       requestSchema.json
     }
-    conf.set(ParquetReadSupportHelper.SPARK_ROW_REQUESTED_SCHEMA, requestSchemaString)
+    conf.set(ParquetReadSupportWrapper.SPARK_ROW_REQUESTED_SCHEMA, requestSchemaString)
   }
